Move TransmodelAPIConfig to correct package
