+++ src/main/java/org/opentripplanner/standalone/config/sandbox/TransmodelAPIConfig.java
@@ -1,8 +1,9 @@
-package org.opentripplanner.standalone.config;
+package org.opentripplanner.standalone.config.sandbox;
 
 import java.util.Collection;
 import java.util.Set;
 import org.opentripplanner.ext.transmodelapi.TransmodelAPIParameters;
+import org.opentripplanner.standalone.config.NodeAdapter;
 
 /**
  * @see TransmodelAPIParameters for documentation of parameters
